old-server/requirements.txt
fastapi
uvicorn
paho-mqtt
influxdb-client
firebase-admin
